Introducing the Apple Pointer
 - Support vertexAttribIPointer

+++ src/main/resources/assets/flywheel/flywheel/layout/block.vert
@@ -3,7 +3,7 @@
 layout(location = 0) in vec3 _flw_v_pos;
 layout(location = 1) in vec4 _flw_v_color;
 layout(location = 2) in vec2 _flw_v_texCoord;
-layout(location = 3) in vec2 _flw_v_light; // TODO: switch to ivec2
+layout(location = 3) in ivec2 _flw_v_light;
 layout(location = 4) in vec3 _flw_v_normal;
 
 void flw_layoutVertex() {
